vektory: update AnsiString vypis_retezy_s_pouzitelnou_rozteci(double Rz,AnsiString separator=",",AnsiString total_separator//pokud je Rz zadané nulové vrátí hodnotu nula, + NOVĚ: pokud chci vypsat všechny načtené řetězy ze souboru retezy.csv použiji parametr Rz=-1
PL: ve výpisu doporučených pohonu, pokud PL není voláno z PO, jednotky také dle PO
PO: unsigned int kapacitaSG; (původně int), v unit1 se pracuje s int
MyString - uprava MyToDouble nepřepíná nyní automaticky z čárky na tečku, ale z oddělovače na daném systému na tečku

+++ parametry.h
@@ -190,7 +190,7 @@ public:		// User declarations
 	Tinput_state input_state;//stav vstupu CT,RD,DD,K
 	Tinput_clicked_edit input_clicked_edit;//zjisteni na ktery edit bylo kliknuto
   Tinput_clicked_icon input_clicked_icon;//zjisteni na ktery icon bylo kliknuto
-	unsigned short kapacitaSG;//požadavek na rozpad na více stejných S&G objektù
+	unsigned int kapacitaSG;//požadavek na rozpad na více stejných S&G objektù
 	bool existuje_pohon;
 	bool MIMO_ROZMEZI;
 	bool form_zobrazen;//detekuje zda je form aktuálnì zobrazen, slouží proto aby pøi zmìnì combo režim pokud si nastavil uživatel formulaø jinam, aby zùstal nastaven dle uživatele
